refac: #8 split math utils
- Split math utils each in its own file
