+++ src/math/add.ts
@@ -1,0 +1,5 @@
+const add = (arr: number[], n: number): number[] => {
+  return arr.map((el) => el + n);
+};
+
+export { add };
