+++ src/math/mean.ts
@@ -1,0 +1,7 @@
+import { sum } from "./sum";
+
+const mean = (arr: number[]): number => {
+  return sum(arr) / arr.length;
+};
+
+export { mean };
