+++ src/math/mod.ts
@@ -1,0 +1,5 @@
+const mod = (x: number, m: number): number => {
+  return ((x % m) + m) % m;
+};
+
+export { mod };
