+++ src/math/multiply.ts
@@ -1,0 +1,5 @@
+const multiply = (arr: number[], n: number): number[] => {
+  return arr.map((el) => el * n);
+};
+
+export { multiply };
