+++ src/math/pow.ts
@@ -1,0 +1,5 @@
+const pow = (arr: number[], n: number): number[] => {
+  return arr.map((el) => Math.pow(el, n));
+};
+
+export { pow };
