+++ src/math/subtract.ts
@@ -1,0 +1,7 @@
+import { add } from "./add";
+
+const subtract = (arr: number[], n: number): number[] => {
+  return add(arr, -n);
+};
+
+export { subtract };
