+++ src/math/sum.ts
@@ -1,0 +1,5 @@
+const sum = (arr: number[]): number => {
+  return arr.reduce((acc, el) => acc + el, 0);
+};
+
+export { sum };
